% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prec_matrix.R
\name{calc_prec_matrix}
\alias{calc_prec_matrix}
\title{Compute n by n covariance matrix W over n grid cells}
\usage{
calc_prec_matrix(neighborhood_matrix, gamma)
}
\arguments{
\item{neighborhood_matrix}{an n by n neighborhood matrix for n grid cells}

\item{gamma}{a real valued parameter in covariance matrix W}
}
\value{
an n by n covariance  matrix
}
\description{
This function computes the covariance matrix of the ensemble 
members for the measured variable, which is a fixed diagonal 
matrix with diagonal elements equal to the empirical variance of the 
corresponding ensemble memebers. Further work will allow W to be a more 
general matrix (e.g., based on a GMRF) with unknown parameters
}
\examples{
n <- 30  # number of grid cells over the globe
neighborhood_matrix <- Matrix::sparseMatrix(1:n, 1:n, x=rep(1,n))
gamma <- 0.1
W <- calc_prec_matrix(neighborhood_matrix, gamma)
Matrix::image(W, xlab="", ylab="", cex=0.5, cex.lab=0.6)
}
\author{
Pulong Ma <mpulong@gmail.com>
}
\keyword{models}

