% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabayes.R
\name{coda_to_df}
\alias{coda_to_df}
\title{Convert JAGS output to Tibble}
\usage{
coda_to_df(coda_object, ...)
}
\arguments{
\item{coda_object}{An object of class \code{mcmc.list} produced by \code{rjags::coda.samples()}.}

\item{...}{Other arguments to pass down, currently unimplemented.}
}
\description{
Converts samples created by \code{rjags::coda.samples()} to a data frame similar to that
produced by the \code{tidybayes} package.
}
