% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_ADtheta_update.R
\name{DA_ADtheta_update}
\alias{DA_ADtheta_update}
\title{Adaptively adjust mean and covariance matrix in the
 proposal distributions}
\usage{
DA_ADtheta_update(theta, ADtheta, oldits)
}
\arguments{
\item{theta}{a list of 3 continuous parameters updated in the MCMC algorithm}

\item{ADtheta}{a list of 6 parameters in the corresponding proposal
distributions for parameters updated in the MCMC algorithm}

\item{oldits}{an integer of the index of the MCMC iteration}
}
\value{
a list of updated parameters in proposal distributions
}
\description{
The function modifies the parameters in proposal distributions for 
corresponding parameters at each iteration to obtain better mixing propertities
in the MCMC algorithm.
}
\author{
Pulong Ma <mpulong@gmail.com>
}
\keyword{models}

