% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DA_theta_MH.R
\name{DA_theta_MH}
\alias{DA_theta_MH}
\title{Update parameters in the MCMC algorithm}
\usage{
DA_theta_MH(outW, theta, prior_theta, ADtheta, n, N, Lj, tol = 1e-10)
}
\arguments{
\item{outW}{a 7 by 1 list containing precomputed quantities
associated with W from the output of function computeW(...)}

\item{theta}{a list of 4 elements containing parameters in the MCMC algorithm}

\item{prior_theta}{parameters in prior distributions}

\item{ADtheta}{a 6 by 1 list containing quantities to adjust mean and 
covariance in proposal distributions}

\item{n}{number of grid cells over the globe}

\item{N}{number of ensemble members}

\item{Lj}{an m by 1 vector containing the number of runs for each 
forcing scenario}

\item{tol}{a very small value to avoid numerical instability}
}
\value{
a list of 4 elements containing updated parameters in the MCMC
iteration
}
\description{
This function updates parameters (gamma, beta, sigma, lambda)  
for the Bayesian statistical model. gamma is updated with a 
discrete uniform prior; beta is updated with the noninformative prior 1
 and random-walk proposal; log of sigma (variance in W)
is updated with the normal prior and random-walk proposal; 
lambda is updated with the normal prior and random-walk proposal.
}
\author{
Pulong Ma <mpulong@gmail.com>
}
\seealso{
DA_GIBBS function
}
\keyword{models}

