% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnnormpdf.R
\name{lnnormpdf}
\alias{lnnormpdf}
\title{Compute the log of pdf of normal random variable with mean u and cariance matrix sigma}
\usage{
lnnormpdf(x, u, sigma)
}
\arguments{
\item{x}{normal random variable (vector)}

\item{u}{mean vector of x}

\item{sigma}{covariance matrix of x}
}
\value{
a real number representing the log of pdf for x ~ N(u, sigma)
}
\description{
This function computes logorithm of the density function for 
normal distribution with mean and cariance matrix
}
\examples{
set.seed(1234)
x <- matrix(rnorm(10), nrow=10, ncol=1)
u <- matrix(rnorm(10), nrow=10, ncol=1)
sigma <- diag(seq(1:10))
out <- lnnormpdf(x, u, sigma)
}
\author{
Pulong Ma <mpulong@gmail.com>
}
\keyword{models}

