% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parDA_GIBBS.R
\name{parDA_GIBBS}
\alias{parDA_GIBBS}
\title{Gibbs sampling with Metropolis-Hastings algorithm for the Bayesian model
of detection and attribution problems}
\usage{
parDA_GIBBS(outW, theta, prior_theta, ADtheta, N, Lj, niter = 20000,
  start_adapting = 500)
}
\arguments{
\item{outW}{a 7 by 1 list containing precomputed quantities associated with W from 
the output of function computeW(...)}

\item{theta}{a list of 3 elements contained in parameters in the MCMC algorithm}

\item{prior_theta}{a list of 2 elements, each of whcih is the quantity 
in the prior distribution of lambda}

\item{ADtheta}{a list of 6 elements, each of which is the parameter in
the corresponding proposal distribution}

\item{N}{an integer representing the number of ensemble members}

\item{Lj}{an m by 1 vector containing the number of runs for each 
forcing scenario}

\item{niter}{an integer specifying the total number of MCMC iterations}

\item{start_adapting}{an integer specifiying when to adapt proposal in 
the MCMC algorithm}
}
\value{
a list of 6 elements containing posterior quantities of 
        parameters, log-likelihood, chisq statistics, and prior:  
        
beta: a matrix holds the posterior samples for the parameter beta with each 
             row corresponding to each beta
             
logsigma: a vector holds the posterior samples for the parameter log of sigma

lambda: a vector holds the posterior samples for the parameter lambda      
   
loglik: a vector holds the log-likelihood evaluated with updated parameters

chisq: a vector holds chisquare statistics for residual consistency test

prior: a vector holds the prior density evaluated with updated parameters
}
\description{
This function implements Gibbs sampling with Metropolis-Hasting
algorithm to sample from posterior distributions for the proposed Bayesian 
statistical model. This function is used in parallel computing of MCMC algorithm
}
\author{
Pulong Ma <mpulong@gmail.com>
}
\seealso{
DA_GIBBS
}
\keyword{models}

