% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parDA_loglik.R
\name{parDA_loglik}
\alias{parDA_loglik}
\title{Compute the loglikelihood function for parallel computing}
\usage{
parDA_loglik(outW, theta, N, Lj)
}
\arguments{
\item{outW}{a 7 by 1 list containing precomputed quantities from 
function parcomputeW(...)}

\item{theta}{a list of 3 elements containing parameters in the MCMC algorithm}

\item{N}{an integer representing the number of ensemble members}

\item{Lj}{an m by 1 vector containing the number of runs for each forcing scenario}
}
\value{
a real number
}
\description{
This function evaluates the likelihood function for the Bayesian
statistical model, and it is used in parallel computing of the MCMC algorithm.
}
\author{
Pulong Ma <mpulong@gmail.com>
}
\seealso{
update_loglik
}
\keyword{models}

