% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMCMC.R
\name{plotMCMC}
\alias{plotMCMC}
\title{Plot results from posterior samples}
\usage{
plotMCMC(MCMCoutputs, N, n, dir = getwd(), burnin = 2000, format = "pdf",
  width = 8, height = 6, weights_plot = TRUE, beta_plot = TRUE,
  beta_box = TRUE, chisq_plot = TRUE)
}
\arguments{
\item{MCMCoutputs}{a list of r elements, where r is the number of EOFs specified, 
and each element is a list of 6 elements containing posterior quantities of 
parameters, log-likelihood, chisq statistics, and prior, which are 
the outputs from MCMC algorithm}

\item{N}{an integer containing the number of ensemble memembers}

\item{n}{an integer containing the number of grid cells}

\item{dir}{a file path specifying where to save the plots}

\item{burnin}{an integer specifying where the posterior samples have good mixing 
propertity}

\item{format}{a character specifying the format of plots to be saved, e.g., "pdf", "png"}

\item{width}{the width of the device (inches for .pdf plot and pixels for others)}

\item{height}{the height of the device (inches for .pdf plot and pixels for others)}

\item{weights_plot}{a logical value specifying whether the posterior plots of
weights are saved}

\item{beta_plot}{a logical value specifying whether the posterior plots of beta
are saved}

\item{beta_box}{a logical value specifying whether the box plots of posterior 
samples for beta are saved}

\item{chisq_plot}{a logical value specifying whether the chisquare test plot 
is saved}
}
\value{
null
}
\description{
This function mainly saves relevant plots for posterior samples.
}
\author{
Pulong Ma <mpulong@gmail.com>
}
\seealso{
parDAbayes
}
\keyword{figures}

