% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simDAdata.R
\name{simDAdata}
\alias{simDAdata}
\title{Simulate data to demonstrate the Bayesian statistical model}
\usage{
simDAdata(n = 30, N = 10, m = 3, Lj = c(5, 3, 7), L0 = 8,
  trend = 30)
}
\arguments{
\item{n}{an integer containing the number of spatial grid cells over the globe}

\item{N}{an integer containing number of ensemble members for measured variable}

\item{m}{an integer containing the number of forcing scenarios}

\item{Lj}{a vector containing the number of model control runs under each forcing
scenarios}

\item{L0}{an integer containing the number of model control runs without any
external forcing}

\item{trend}{trend of the simulated data for the ensemble of the measured variable
with default value 30}
}
\value{
a list of 3 elements containing an n by N matrix of the ensemble of the
measured variable, an n by m matrix of model outputs for measured variable under
m forcing scenarios, and an n by L0 matrix of model control runs for measured  
variable without any external forcing scenario
}
\description{
This function simulates the ensemble of the measured variable, model 
outputs for the measured variable under different forcing scenarios, and model outputs
for the measured variable without any external forcing scenario.
}
\examples{
n <- 30
N <- 10
m <- 3
Lj <- c(5, 3, 7)
L0 <- 8
trend <- 30
DAdata <- simDAdata(n, N, m, Lj, L0, trend)
# ensembles of the measured variable
y <- DAdata[[1]]
# model outputs for the measured variable under different forcing scenarios
x <- DAdata[[2]]
# model outputs for the measured variable without any external forcing scenario
x0 <- DAdata[[3]]
}
\author{
Pulong Ma <mpulong@gmail.com>
}
\keyword{models}

