% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_loglik.R
\name{update_loglik}
\alias{update_loglik}
\title{Compute loglikelihood function in the model}
\usage{
update_loglik(outW, theta, n, N, Lj)
}
\arguments{
\item{outW}{a 7 by 1 list containing precomputed quantities associated with W from 
the output of function computeW(...)}

\item{theta}{a list of 5 elements containing parameters in MCMC algorithm}

\item{n}{an integer representing the number of grid cells over the globe}

\item{N}{an integer representing the number of ensemble members}

\item{Lj}{an m by 1 vector containing the number of runs for each 
forcing scenario}
}
\value{
a real rumber
}
\description{
This function computes the logarithm of the likelihood function 
for the Bayesian statistical model.
}
\author{
Pulong Ma <mpulong@gmail.com>
}
\keyword{models}

