% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_var.R
\name{update_var}
\alias{update_var}
\title{Adjust quantities in proposal distributions}
\usage{
update_var(newvec, oldcov, oldmean, oldits)
}
\arguments{
\item{newvec}{updated parameters in prior distribution}

\item{oldcov}{covariance matrix in proposal distribution}

\item{oldmean}{mean in proposal distribution}

\item{oldits}{index of the previous MCMC iteration}
}
\value{
updated quantities in proposal distributions
}
\description{
This function specifies how we adjust mean and covariance matrix
in the proposal distribution
}
\author{
Pulong Ma <mpulong@gmail.com>
}
\keyword{models}

