% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lnnormpdf.R
\name{lnnormpdf}
\alias{lnnormpdf}
\title{Compute the logarithm of the pdf of normal random variable with mean u and cariance matrix sigma}
\usage{
lnnormpdf(x, u, sigma)
}
\arguments{
\item{x}{normal random variable (vector)}

\item{u}{mean vector of x}

\item{sigma}{covariance matrix of x}
}
\value{
a real number representing the logarithm of pdf for x ~ N(u, sigma)
}
\description{
This function computes the logarithm of the density function for 
normal distribution with mean and covariance matrix
}
\examples{
set.seed(1234)
n <- 10
x <- matrix(rnorm(n), nrow=n, ncol=1)
u <- matrix(rnorm(n), nrow=n, ncol=1)
sigma <- exp(-as.matrix(dist(rnorm(n))))
out <- lnnormpdf(x, u, sigma)
}
\author{
Pulong Ma <mpulong@gmail.com>
}
\keyword{models}

